ap-4324: add type to opponents table
Create individual and organisation opponent subclasses
