+++ app/models/application_merits_task/opponent/individual.rb
@@ -1,0 +1,9 @@
+module ApplicationMeritsTask
+  module Opponent
+    class Individual < BaseOpponent
+      def full_name
+        "#{first_name} #{last_name}".strip
+      end
+    end
+  end
+end
