+++ app/models/application_merits_task/opponent/organisation.rb
@@ -1,0 +1,6 @@
+module ApplicationMeritsTask
+  module Opponent
+    class Organisation < BaseOpponent
+    end
+  end
+end
