+++ app/models/legal_aid_application.rb
@@ -27,7 +27,7 @@ class LegalAidApplication < ApplicationRecord
   has_one :statement_of_case, class_name: "ApplicationMeritsTask::StatementOfCase", dependent: :destroy
   has_one :gateway_evidence, dependent: :destroy
   has_one :uploaded_evidence_collection, dependent: :destroy
-  has_many :opponents, class_name: "ApplicationMeritsTask::Opponent", dependent: :destroy
+  has_many :opponents, class_name: "ApplicationMeritsTask::Opponent::BaseOpponent", dependent: :destroy
   has_one :domestic_abuse_summary, class_name: "ApplicationMeritsTask::DomesticAbuseSummary", dependent: :destroy
   has_one :parties_mental_capacity, class_name: "ApplicationMeritsTask::PartiesMentalCapacity", dependent: :destroy
   has_one :latest_incident, -> { order(occurred_on: :desc) }, class_name: "ApplicationMeritsTask::Incident", inverse_of: :legal_aid_application, dependent: :destroy
