+++ db/migrate/20230717071401_add_type_opponents.rb
@@ -1,0 +1,7 @@
+class AddTypeOpponents < ActiveRecord::Migration[7.0]
+  def change
+    add_column :opponents, :type, :string, default: "ApplicationMeritsTask::Opponent::Individual"
+    add_column :opponents, :organisation_name, :string
+    add_column :opponents, :organisation_type, :string
+  end
+end
